Removed user defaults saving, made constants private, moved is valid email check method to separated class
