+++ ShopClient/Constants/Constants.swift
@@ -49,5 +49,4 @@ struct SessionData {
     static let accessToken = "AccessToken"
     static let expiryDate = "ExpiryDate"
     static let email = "Email"
-    static let userDefaultsLoggedIn = "UserLoggedIn"
 }
